Update LibMath.cpp
Update LibMath.cpp

+++ ProjMath/LibMath.cpp
@@ -20,7 +20,7 @@ int Multiplicar(int num1,int num2)
 
 float Dividir(int num1,int num2)
 {
-    return num1/num2;
+    return (float)num1/(float)num2;
 }
 
 float Potencia(int num1,int num2)
